requirements.txt
fastapi
uvicorn
langchain
pypdf
openai
faiss-cpu
python-multipart
python-dotenv
tiktoken